Update README.md
# NAIJACONNECT-A-NIGERIAN-ONLINE-MARKET-PLACE

This is my Mini project assessment for my second 3MTT Project on Darey.io: Responsive landing page. In this project 

I created an online Marketplace where artisans can be hired to provide services to households and companies in need of their services. 

In line with the objective of this project, I designed a responsive landing page for my product (NaijaConnect) using html and CSS elements only.

The page contains element such as header which include the title of the page, a logo image of my product (NaijaConnect) as well as a navigation bar. 

The body of the web page contains different sections which includes Features, How it works, as well as a call to action (cto) button of download. The feature or hero sections gives an introduction of the specification of NaijaConnect as well as it importance as a service based platform. Other elements of html such as paragraphs and strong was used to create emphasis. Same element applied to the How it works section as it was utilized in the other sections of the web page. The webpage also contaion a footer section as directed by the task instruction. this is aimed at meeting all the requirement of the project. 

To ensure that the page is visually appealing, mobile-friendly and structured for a grreat user experience, CSS was used to achieve is aesthetics and to also ensure it is when aligned and structured while working on a mobile device.

Also, The body of the web page contains small images which are used to describe the key feature as well as the function which the product performs like; Local discover, Verified providers and real reviews.

The file is hereby attached in a well structured and mobile friendly  fashion. Thank you!